Add option for HOST to config/app.js
Binding the host to 'localhost' is not always ideal.  For example, if placing your application inside Docker, you need to bind to 0.0.0.0.  Or perhaps you're on a physical machine, and need to bind to a specific IP address (192.168.0.1).

This code adds the configuration option.  

However, there is a another modification required.  There must be a corresponding code change in the supercharge foundation.  
     ./framework/src/foundation/http/kernel.js

I'll submit that pull request also.